fix(kiosk): rename PiNfcReader -> Acr1552uReader
Real hardware is an ACS ACR1552U USB NFC reader (CCID/PCSC), not a
PN532 GPIO HAT. Backend will use pyscard via the PC/SC daemon.
